# swagger.yaml
openapi: 3.0.0
info:
  title: Urban Farming Platform API
  description: API for Interactive Urban Farming Platform
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
...
    RentalSpace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        pricePerMonth:
          type: number
        availability:
          type: boolean
...
paths:
...
  /rentals:
    get:
      tags:
        - Rentals
      summary: Get rental spaces
      parameters:
        - name: location
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of rental spaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RentalSpace'
  /rentals/{id}/book:
    post:
      tags:
        - Rentals
      summary: Book a rental space
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - startDate
                - endDate
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ==================== ORDERS ====================
...